Updates the tests
